+++ source/main/commands/file-rename.js
@@ -14,9 +14,9 @@
 
 const path = require('path')
 const ZettlrCommand = require('./zettlr-command')
-const ignoreFile = require('../../common/util/ignore-file')
 const sanitize = require('sanitize-filename')
-const hash = require('../../common/util/hash')
+const isFile = require('../../common/util/is-file')
+const ALLOWED_FILETYPES = require('../../common/data.json').filetypes
 
 class FileRename extends ZettlrCommand {
   constructor (app) {
@@ -34,32 +34,31 @@ class FileRename extends ZettlrCommand {
     // We need to prepare the name to be correct for
     // accurate checking whether or not the file
     // already exists
-    arg.name = sanitize(arg.name, { replacement: '-' })
-    // Make sure we got an extension.
-    if (ignoreFile(arg.name)) arg.name += '.md'
+    let newName = sanitize(arg.name, { replacement: '-' })
+
+    // If no valid filename is provided, assume .md
+    let ext = path.extname(newName).toLowerCase()
+    if (!ALLOWED_FILETYPES.includes(ext)) {
+      newName += '.md'
+    }
 
     let file = this._app.findFile(arg.hash)
-    if (!file) return global.log.error(`Could not find file ${arg.hash}`)
+    if (file === null) {
+      return global.log.error(`Could not find file ${arg.hash}`)
+    }
 
     // Test if we are about to override a file
-    if (this._app.findFile(hash(path.join(file.dir, arg.name)))) {
+    if (isFile(path.join(file.dir, newName))) {
       // Ask for override
-      let result = await this._app.getWindow().askOverwriteFile(arg.name)
+      let result = await this._app.getWindow().askOverwriteFile(newName)
       if (result.response === 0) return // No override wanted
     }
 
-    // askOverwriteFile
     try {
-      await this._app.getFileSystem().runAction('rename-file', {
-        'source': file,
-        'info': { 'name': arg.name }
-      })
+      await this._app.getFileSystem().renameFile(file, newName)
     } catch (e) {
       global.log.error('Error during renaming file: ' + e.message, e)
     }
-
-    // And done. FSAL has already notified the app of a necessary update
-    // global.application.fileUpdate(arg.hash, file.hash)
     return true
   }
 }